Refactoring: Extracted 'UART Communication' tab into a separate user control to improve code readability and loose coupling between independent software components
