+++ PS5 NOR Modifier/Form1.resx
@@ -69,13 +69,4 @@ coffee addiction! Click here to donate to me. Even a dollar helps :) Don't forge
   <metadata name="statusStrip1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
     <value>17, 17</value>
   </metadata>
-  <data name="label25.Text" xml:space="preserve">
-    <value>You can send custom commands
-with UART using this feature. Note
-that any command entered here
-will return unparsed results. Please
-also be sure to know what you are
-doing as sending unsafe commands
-can damage your device.</value>
-  </data>
 </root>
