+++ PS5 NOR Modifier/PS5 NOR Modifier.csproj.user
@@ -7,5 +7,8 @@
     <Compile Update="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Update="UserControls\UART\UartUserControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
   </ItemGroup>
 </Project>
